Moved analysis to a "passes" module
